refactor(api): move security-pre-handler into organization-entities
