+++ api/tests/shared/unit/application/security-pre-handlers_test.js
@@ -630,176 +630,6 @@ describe('Shared | Unit | Application | SecurityPreHandlers', function () {
     });
   });
 
-  describe('#checkCertificationCenterIsNotScoManagingStudents', function () {
-    let checkOrganizationIsScoAndManagingStudentUsecaseStub;
-    let centerRepositoryStub;
-
-    let dependencies;
-
-    beforeEach(function () {
-      checkOrganizationIsScoAndManagingStudentUsecaseStub = { execute: sinon.stub() };
-      centerRepositoryStub = {
-        findActiveScoOrganizationId: sinon.stub(),
-      };
-
-      dependencies = {
-        checkOrganizationIsScoAndManagingStudentUsecase: checkOrganizationIsScoAndManagingStudentUsecaseStub,
-        centerRepository: centerRepositoryStub,
-      };
-    });
-
-    context('Successful cases', function () {
-      context('when certification center does not belong to an active sco organization', function () {
-        it('should authorize access to resource when the user is authenticated', async function () {
-          // given
-          const request = {
-            auth: {
-              credentials: {
-                accessToken: 'valid.access.token',
-                userId: 1234,
-              },
-            },
-            params: {
-              certificationCenterId: 5678,
-            },
-          };
-          dependencies.checkOrganizationIsScoAndManagingStudentUsecase.execute.resolves(false);
-          dependencies.centerRepository.findActiveScoOrganizationId.resolves(null);
-
-          // when
-          const response = await securityPreHandlers.checkCertificationCenterIsNotScoManagingStudents(
-            request,
-            hFake,
-            dependencies,
-          );
-
-          // then
-          expect(response.source).to.be.true;
-        });
-      });
-
-      context('when certification center id is in request params', function () {
-        it('should authorize access to resource when the user is authenticated, member of certification center and the organization associated is not SCO managing students', async function () {
-          // given
-          const request = {
-            auth: {
-              credentials: {
-                accessToken: 'valid.access.token',
-                userId: 1234,
-              },
-            },
-            params: {
-              certificationCenterId: 5678,
-            },
-          };
-          dependencies.checkOrganizationIsScoAndManagingStudentUsecase.execute.resolves(false);
-          dependencies.centerRepository.findActiveScoOrganizationId.resolves(1);
-
-          // when
-          const response = await securityPreHandlers.checkCertificationCenterIsNotScoManagingStudents(
-            request,
-            hFake,
-            dependencies,
-          );
-
-          // then
-          expect(response.source).to.be.true;
-        });
-      });
-
-      context('when certification center id is in request payload', function () {
-        it('should authorize access to resource when the user is authenticated, member of certification center and the organization associated is not SCO managing students', async function () {
-          // given
-          const request = {
-            auth: {
-              credentials: {
-                accessToken: 'valid.access.token',
-                userId: 1234,
-              },
-            },
-            payload: {
-              data: {
-                attributes: {
-                  certificationCenterId: 5678,
-                },
-              },
-            },
-          };
-          dependencies.checkOrganizationIsScoAndManagingStudentUsecase.execute.resolves(false);
-          dependencies.centerRepository.findActiveScoOrganizationId.resolves(1);
-
-          // when
-          const response = await securityPreHandlers.checkCertificationCenterIsNotScoManagingStudents(
-            request,
-            hFake,
-            dependencies,
-          );
-
-          // then
-          expect(response.source).to.be.true;
-        });
-      });
-    });
-
-    context('Error cases', function () {
-      it('should forbid resource access when user was not previously authenticated', async function () {
-        // given
-        const request = {
-          payload: {
-            data: {
-              attributes: {
-                certificationCenterId: 5678,
-              },
-            },
-          },
-        };
-
-        // when
-        const response = await securityPreHandlers.checkCertificationCenterIsNotScoManagingStudents(
-          request,
-          hFake,
-          dependencies,
-        );
-
-        // then
-        expect(response.statusCode).to.equal(403);
-        expect(response.isTakeOver).to.be.true;
-      });
-
-      it('should forbid resource access when the certification center does belong to SCO Organization and manage students', async function () {
-        // given
-        const request = {
-          auth: {
-            credentials: {
-              accessToken: 'valid.access.token',
-              userId: 1234,
-            },
-          },
-          payload: {
-            data: {
-              attributes: {
-                certificationCenterId: 5678,
-              },
-            },
-          },
-        };
-        dependencies.checkOrganizationIsScoAndManagingStudentUsecase.execute.resolves(true);
-        dependencies.centerRepository.findActiveScoOrganizationId.resolves(1);
-
-        // when
-        const response = await securityPreHandlers.checkCertificationCenterIsNotScoManagingStudents(
-          request,
-          hFake,
-          dependencies,
-        );
-
-        // then
-        expect(response.statusCode).to.equal(403);
-        expect(response.isTakeOver).to.be.true;
-      });
-    });
-  });
-
   describe('#hasAtLeastOneAccessOf', function () {
     let belongsToOrganizationStub;
     let hasRoleSuperAdminStub;
